backend/requirements.txt
requests
beautifulsoup4
python-dotenv
llama-index
flask
flask-cors
streamlit